PHOENIX-6246 : Remove flakes from PointInTimeQueryIT
